Restructure backend graphql due to import issues
The previous structure lended itself to circular import issues. This
also looks closer to the front-end.
