+++ fittrak/workouts/graphql/query.py
@@ -4,14 +4,24 @@ import graphene
 
 from django.db.models import Count
 
-from workouts.models import Exercise
-from workouts.models import ExerciseType as ExerciseTypeModel
-from workouts.models import MuscleGroup, Set, Workout
+from graphql import GraphQLError
 
-from .exercises import ExerciseType, ExerciseTypeType
-from .helpers import get_object
-from .sets import SetType
-from .workouts import MuscleGroupType, WorkoutStatusesEnum, WorkoutType
+from workouts.models import (
+    Exercise,
+    ExerciseType as ExerciseTypeModel,
+    Set,
+    Workout,
+    MuscleGroup,
+)
+
+from .types import (
+    ExerciseType,
+    ExerciseTypeType,
+    SetType,
+    MuscleGroupType,
+    WorkoutStatusesEnum,
+    WorkoutType,
+)
 
 
 class Query:
@@ -92,10 +102,20 @@ class Query:
     def resolve_workout(_, info, workout_id):
         user = info.context.user
 
-        return get_object(Workout, {"id": workout_id, "user": user.id})
+        try:
+            workout = Workout.objects.get(is_active=True, id=workout_id, user=user_id)
+        except Workout.DoesNotExist:
+            raise GraphQLError("Workout not found.")
+
+        return workout
 
     @staticmethod
     def resolve_exercise(_, info, exercise_id):
         user = info.context.user
 
-        return get_object(Exercise, {"id": exercise_id, "user": user.id})
+        try:
+            exercise = Exercise.objects.get(
+                is_active=True, id=exercise_id, user=user.id
+            )
+        except Exercise.DoesNotExist:
+            raise GraphQLError("Exercise not found.")
